moved the content of cmethods.py to __init__.py and adjusted this in all places
