+++ docs/src/getting_started.rst
@@ -23,7 +23,7 @@ method specific documentation.
     :caption: Apply the Linear Scaling bias correction technique on 1-dimensional data
 
     import xarray as xr
-    from cmethods.CMethods import CMethods as cm
+    from cmethods import CMethods as cm
 
     obsh = xr.open_dataset('input_data/observations.nc')
     simh = xr.open_dataset('input_data/control.nc')
@@ -41,7 +41,7 @@ method specific documentation.
     :caption: Apply the Quantile Delta Mapping bias correction technique on 1-dimensional data
 
     import xarray as xr
-    from cmethods.CMethods import CMethods as cm
+    from cmethods import CMethods as cm
 
     obsh = xr.open_dataset('input_data/observations.nc')
     simh = xr.open_dataset('input_data/control.nc')
